Refactoring: Split Large Class using Extract class The previously monolithic ChessGame god class was refactored into multiple cohesive classes using the Extract Class and Extract Subclass refactoring techniques.
Each logical responsibility was isolated into its own file:
- `ChessGame.cs` – Now contains only core game state management and delegates responsibilities.
- `BoardInitializer.cs` – Responsible for the initial piece placement and king positioning.
- `BoardEvaluator.cs` – Encapsulates logic to check if a square is under attack.
- `MoveValidator.cs` – Handles all move legality checks and simulation for check resolution.
- `GameLogger.cs` – Logs game result messages.
- `GameStateChangedEventArgs.cs` – Clean event argument model to notify listeners of game state changes.

`Improved Modularity`: Each class now has a single responsibility. This makes the system easier to reason about and reduces the risk of unintended side effects when changing logic in one component.
`Simplified Maintenance`: Developers can now locate and update specific logic (e.g., move validation or board setup) without navigating a large file or risking unrelated code.
